transfer of all css in tsx files to separate css files, all css files now are in css folder, minor file changes
