+++ .gitignore
@@ -1,1 +1,4 @@
 .idea/
+frontend/node_modules/
+frontend/src/.DS_Store
+package-lock.json
