+++ frontend/src/components/index.tsx
@@ -1,3 +1,3 @@
-export {Navigation} from './Navigation';
-export {PrivateRoute} from './PrivateRoute';
-export {View} from './View';
+export { Navigation } from './Navigation';
+export { PrivateRoute } from './PrivateRoute';
+export { View } from './View';
