+++ frontend/src/css/LoginPage.css
@@ -1,3 +1,4 @@
 .login-background {
+  min-height: 100vh;
   background-color: lightblue;
 }
