+++ frontend/src/css/Navigation.css
@@ -1,0 +1,10 @@
+.header {
+  position: fixed;
+  z-index: 999;
+  width: 100%;
+}
+
+.logo {
+  height: 60px;
+  width: 120px;
+}
