+++ frontend/src/css/View.css
@@ -1,0 +1,17 @@
+.site-layout {
+  padding: 0 50px;
+  margin-top: 64px;
+}
+
+.breadcrumb {
+  margin: 16px 0;
+}
+
+.site-layout-background {
+  padding: 24px;
+  min-height: 380px;
+}
+
+.footer {
+  text-align: center;
+}
