+++ frontend/src/index.tsx
@@ -1,19 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import './index.css';
-import {CalendarView, HomeView, LoginPage} from './views';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { CalendarView, HomeView, LoginPage } from './views';
 import reportWebVitals from './reportWebVitals';
-import {ProvideAuth} from './utils/auth';
+import { ProvideAuth } from './utils/auth';
 import { PrivateRoute } from './components';
 
 ReactDOM.render(
   <ProvideAuth>
     <Router>
       <Switch>
-        <PrivateRoute path='/calendar' component={CalendarView} />
-        <PrivateRoute path='/home' component={HomeView} />
-        <Route exact path='/' component={LoginPage} />
+        <PrivateRoute path='/calendar' component={ CalendarView } />
+        <PrivateRoute path='/home' component={ HomeView } />
+        <Route exact path='/' component={ LoginPage } />
       </Switch>
     </Router>
   </ProvideAuth>,
