--- frontend/src/views/CalendarView.css
@@ -1,25 +1,0 @@
-.logo {
-  display:inline-block;
-  width:8% !important;
-  margin-right: 20px;
-  margin-left: -20px;
-}
-
-#components-layout-demo-fixed .logo {
-  float: left;
-  width: 120px;
-  height: 31px;
-  margin: 16px 24px 16px 0;
-  background: rgba(255, 255, 255, 0.2);
-}
-.site-layout .site-layout-background {
-  background: #fff;
-}
-
-.buttons {
-  text-align: center;
-  display:inline-block;
-  vertical-align:top;
-  padding: 0px;
-  margin: 0px;
-}
