+++ frontend/src/views/HomeView.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
-import {Table } from 'antd';
-import 'antd/dist/antd.css';
-import './HomeView.css';
-import { View} from '../components';
+import { Table } from 'antd';
+import { View } from '../components';
 import { columns, data } from './temporaryData';
-
-
+import 'antd/dist/antd.css';
+import '../css/HomeView.css';
 
 export const HomeView = () => {
   return (
     <View>
-      <Table columns={columns} dataSource={data} className="table" />
+      <Table columns={columns} dataSource={data} className="table"/>
     </View>
   );
 };
