+++ frontend/src/views/index.tsx
@@ -1,3 +1,3 @@
-export {CalendarView} from './CalendarView';
-export {HomeView} from './HomeView';
-export {LoginPage} from './LoginPage';
+export { CalendarView } from './CalendarView';
+export { HomeView } from './HomeView';
+export { LoginPage } from './LoginPage';
